Split swarm callback module
